requirements.txt
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
google-api-core==1.22.1
google-auth==1.21.0
google-cloud-vision==1.0.0      
googleapis-common-protos==1.52.0
grpcio==1.31.0
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyperclip==1.8.0
PySimpleGUI==4.29.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1